+++ src/test/resources/MockitoTest.groovy
@@ -15,6 +15,9 @@ class MockitoTest extends Specification {
 		mockedList.get(3) >> {
 			throw new IndexOutOfBoundsException('out of bound')
 		}
+		mockedList.get(4) >> {
+			throw new IndexOutOfBoundsException()
+		}
 		mockedList.get(0) >> 'some'
 		personDao.save({ Object argument ->
 			return 'Spock'.equalsIgnoreCase(((Person)argument).getName())
@@ -34,7 +37,7 @@ class MockitoTest extends Specification {
 		1 * mockedList.get(_)
 		1 * mocked.someMethod(_ as Boolean,_ as Byte,_ as Character,_ as Integer,_ as Long,_ as Float,_ as Double)
 		1 * mocked.someOtherMethod(_ as Short,_ as String,_ as List,_ as Set,_ as Map,_ as Collection,_ as Iterable)
-		1 * mocked.someOtherOtherMethod(_,_ as String,_ as String,'some string',null,!null,!null,_ as List<Object>,_ as Set<Object>,_ as Collection<Object>,_ as Iterable<Object>,_ as Map<Long,List>)
+		1 * mocked.someOtherOtherMethod(_,_ as String,_ as String,null,!null,!null,_ as List<Object>,_ as Set<Object>,_ as Collection<Object>,_ as Map<Long,List>)
 		1 * mocked.someOtherOtherOtherMethod({
 			it.startsWith('prefix')
 		} as String)